Use unsigned long to hold 32-bit values, for portability

+++ breath/cpu/example/cpu.cpp
@@ -38,12 +38,12 @@ processor_brand_string()
 
     // Reference: Intel Manual, 3-176, Vol. 2A
     //
-    unsigned const      from = 0x8000'0002 ;
-    unsigned const      to   = 0x8000'0004 ;
+    unsigned long const from = 0x8000'0002 ;
+    unsigned long const to   = 0x8000'0004 ;
 
     std::size_t const   len = 16 * ( to - from + 1 ) ;
     char                s[ len ] ;
-    for ( unsigned i = from ; i <= to ; ++ i ) {
+    for ( unsigned long i = from ; i <= to ; ++ i ) {
         cpuid_result const  info = get_cpuid_info( i, 0 ) ;
         char * const        p = s + 16 * ( i - from ) ;
         std::memcpy( p,      &info.eax, sizeof( info.eax ) ) ;
